Finished writing and testing the Battle class.
Most of the combat system and character progression is finished. Only need to tweak a couple of things. The Battle class has a bug where it is not providing the correct values to the receive attack function. Exp and leveling up needs to fleshed out as well as item drops. Need to fix an issue where the enemy's death message is not being shown. Other than that I just need to write the GraphBuilder class, the Game class, and then finally design all of the Rooms, Enemies, Items, and put it all together and test it. Hopefully I have enough time tomorrow because I am done for the night.