did some testing
created a player, items, enemies, and a bunch of nodes to do some testing with all of those classes. found lots of small issues that need fixed. also discovered a major logical issue with the OneWayNode class and the FallNode class which will require some redesign. Need to iron out all of these issues, and then build the GraphBuilder and Game classes