made some balancing changes
not perfect, but it ist at least beatable in about 10-20 min now. really allows the player to snowball once they level up a couple times and get lucky enough to defeat the executioner and obtain some endgame items. definitely needs changed because it lets the player pretty much just walk over all 3 of the final bosses and then oneshot grool at the end if they have all of the items. but at least it is beatable. earlygame = hard, lategame = easy